feat(670): widen TreeNode.parent_path u128 -> TreePath [u32;8]
The 16-bit-per-level u128 packing corrupted ancestor bits for any
token id >= 65,536 and could never recover them (Bonsai vocab =
248,320 needs 18-bit ids). TreePath keeps the 32-byte Copy footprint,
derives Eq/Ord/Hash (lexicographic slot order == old packed order for
same-length paths), and replaces the shift/mask idioms with
root/push(token, depth)/parent(depth)/token_at(k).

- katgpt-core: TreePath + 4 unit tests (legacy-parity at <=65,536,
  Bonsai-vocab round-trip incl. 248,319); gdn_tree_verify parent keys
- katgpt-speculative: tree_builder (~50 sites), dd_tree extract fns,
  lodestar, ilc (lossy path hash preserved bit-identical at <=65,536),
  bigram_markov acceptance helper, tests
- katgpt-forward/attn/pruners + root: consumer sites
- extract_parent_tokens* now take TreePath (coordinated API break)

GOAT: G1 existing suites green (329+124+25+126+1903+201); G2 round-trip
at 248,320 pinned; G3 same-size Copy, bench_424 verdict unchanged;
G4 = riir-ai 717 gate invariants promoted clean16 -> all-positions
(lands with the riir-ai commit). Issue 670 resolved.

+++ crates/katgpt-speculative/src/bigram_markov.rs
@@ -933,15 +933,16 @@ and in what order the agents took their turns upon the ground.";
     /// speculative decoder realises, because the verifier checks every path
     /// in one batched forward and commits the longest match.
     ///
-    /// A node's `parent_path` packs its whole ancestor chain including itself,
-    /// one token per level, depth 0 at level 0 (root first).
+    /// A node's `parent_path` holds its whole ancestor chain including
+    /// itself, one `u32` token per level, root at slot 0
+    /// (`tree_builder`: `node_path = parent_path.push(token, depth)`).
     fn tree_acceptance(tree: &[TreeNode], target: &[u32]) -> usize {
         let mut best = 0usize;
         for n in tree {
             if n.depth + 1 > target.len() || n.depth < best {
                 continue;
             }
-            let matches = (0..=n.depth).all(|k| n.parent_path.token_at(k) as u32 == target[k]);
+            let matches = (0..=n.depth).all(|k| n.parent_path.token_at(k) == target[k]);
             if matches {
                 best = n.depth + 1;
             }
